[server] Minor refactor
